requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
black==22.8.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
fastapi==0.78.0
frozenlist==1.3.1
graphql-core==3.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
jarowinkler==1.0.2
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==6.1.2
rapidfuzz==2.0.11
redis==4.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
strawberry-graphql==0.126.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
yarl==1.8.1
